Reverse compatibility for iOS4 and iOS5
On iOS6 the UIControlEventEditingChanged is not triggered when you
change a textfield.text programatically. It was assumed and
textFieldValueChanged: was not invoked incorrecttly, but in iOS4|5 is
ALWAYS invoked, causing the optionsView to disappear and appear again,
if you try to set the field's text on your own code. The quick solution
is to invoke the hide AFTER the delegate call